Fixing tests and moving OE code to its own project
