+++ src/Microsoft.SqlTools.SqlCore/Metadata/ObjectMetadata.cs
@@ -3,8 +3,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-#nullable disable
-
 namespace Microsoft.SqlTools.ServiceLayer.Metadata.Contracts
 {
     /// <summary>
