+++ src/Microsoft.SqlTools.SqlCore/Microsoft.SqlTools.SqlCore.csproj
@@ -1,0 +1,38 @@
+<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<Nullable>disable</Nullable>
+		<EnableDefaultItems>false</EnableDefaultItems>
+		<EnableDefaultCompileItems>false</EnableDefaultCompileItems>
+		<EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
+		<EmbeddedResourceUseDependentUponConvention>false</EmbeddedResourceUseDependentUponConvention>
+		<EnableDefaultNoneItems>false</EnableDefaultNoneItems>
+		<DefineConstants>$(DefineConstants);NETCOREAPP1_0;TRACE</DefineConstants>
+		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+		<PreserveCompilationContext>true</PreserveCompilationContext>
+		<AssemblyTitle>SqlTools Hosting Library</AssemblyTitle>
+		<Description>Provides hosting services for SqlTools applications.</Description>
+	</PropertyGroup>
+	<ItemGroup>
+		<Compile Include="**\*.cs" Exclude="**/obj/**/*.cs" />
+	</ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Newtonsoft.Json" />
+		<PackageReference Include="Microsoft.Extensions.DependencyModel" />
+		<PackageReference Include="System.Runtime.Loader" />
+		<PackageReference Include="System.Composition" />
+		<PackageReference Include="System.Text.Encodings.Web" />
+		<PackageReference Include="Microsoft.SqlServer.SqlManagementObjects" />
+	</ItemGroup>
+	<ItemGroup>
+		<InternalsVisibleTo Include="Microsoft.SqlTools.ServiceLayer.UnitTests" />
+		<InternalsVisibleTo Include="Microsoft.SqlTools.ServiceLayer.IntegrationTests" />
+		<InternalsVisibleTo Include="Microsoft.SqlTools.ServiceLayer.Test.Common" />
+	</ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="../Microsoft.SqlTools.Hosting/Microsoft.SqlTools.Hosting.csproj" />
+	</ItemGroup>
+	<ItemGroup>
+		<EmbeddedResource Include="Localization\*.resx" />
+		<None Include="Localization\sr.strings" />
+	</ItemGroup>
+</Project>
