+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/INodeFilter.cs
