+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodeFilterProperty.cs
