+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodeObservableCollection.cs
