+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodeOrFilter.cs
