+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodePropertyFilter.cs
@@ -3,8 +3,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-#nullable disable
-
 using System;
 using System.Collections.Generic;
 using System.Text;
