+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodeSmoProperty.cs
