+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodeTypes.cs
