+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/NodeTypes.tt
