+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/TreeNode.cs
