+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/Nodes/TreeNodeWithContext.cs
