+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/ObjectExplorerServerInfo.cs
