+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/FolderNode.cs
