+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoCollectionWrapper.cs
