+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoDatabaseCustomNode.cs
