+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoExtensions.cs
