+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoQueryModel.cs
