+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoQueryModel.tt
