+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoQueryModelDefinition.xml
