+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoTreeNodesDefinition.xml
