+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoTriggerCustomNode.cs
