+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoUserCustomNode.cs
