+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SmoWrapper.cs
