+++ src/Microsoft.SqlTools.SqlCore/ObjectExplorer/SmoModel/SqlHistoryTableQuerier.cs
